requirements.txt
automata-lib==9.0.0
blinker==1.9.0
cached_method==0.1.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
frozendict==2.4.6
graphviz==0.20.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.6
pillow==11.2.1
pydot==4.0.0
pyformlang==1.0.11
pyparsing==3.2.3
termcolor==3.1.0
typing_extensions==4.13.2
Werkzeug==3.1.3
